+++ ReleaseNotes.md
@@ -1,3 +1,10 @@
+# v4.8.4-1
+
+## Summary
+This release reduces the number of DB queries issued to the CTA catalogue.
+### Bug Fixes
+- cta/CTA#275 - Avoid DB queries via getTapesByVid in OStoreDB::fetchMountInfo
+
 # v4.8.3-1
 
 ## Summary